data-pipeline/requirements.txt
requests>=2.31.0
pandas>=2.1.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.4.0
